Split form and validator out of controller
