+++ app/controllers/keywords_controller.rb
@@ -1,41 +1,5 @@
 # frozen_string_literal: true
 
-require 'csv'
-
-class CsvValidator < ActiveModel::Validator
-  def validate(record)
-    file = record.instance_variable_get(:@file)
-
-    return record.errors.add(:base, I18n.t('keyword.file_must_be_csv')) unless file.content_type == 'text/csv'
-    return record.errors.add(:base, I18n.t('keyword.keyword_range')) unless CSV.read(file).count.between?(1, 1000)
-  end
-end
-
-class CsvImportForm
-  include ActiveModel::Model
-  include ActiveModel::Validations
-  validates_with CsvValidator
-
-  def initialize(user, file)
-    @user = user
-    @file = file
-  end
-
-  def save
-    return self unless valid?
-
-    bulk_data = []
-
-    CSV.foreach(@file.path) do |row|
-      bulk_data << { user_id: @user.id, keyword: row[0], created_at: Time.current, updated_at: Time.current }
-    end
-
-    # rubocop:disable Rails/SkipsModelValidations
-    Keyword.insert_all(bulk_data, returning: %w[id keyword])
-    # rubocop:enable Rails/SkipsModelValidations
-  end
-end
-
 class KeywordsController < ApplicationController
   before_action :ensure_authentication
